+++ src/Exceptions/GazeBlobExpiredException.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Naoray\GazeLaravel\Exceptions;
 
-class GazeBlobExpiredException extends GazeException
+class GazeBlobExpiredException extends GazeException implements TerminalGazeException
 {
 }