+++ src/Exceptions/GazeRestoreFailedException.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Naoray\GazeLaravel\Exceptions;
 
-class GazeRestoreFailedException extends GazeException
+class GazeRestoreFailedException extends GazeException implements TransientGazeException
 {
 }