+++ src/Exceptions/GazeSanitizeFailedException.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Naoray\GazeLaravel\Exceptions;
 
-class GazeSanitizeFailedException extends GazeException
+class GazeSanitizeFailedException extends GazeException implements TransientGazeException
 {
 }