+++ src/Exceptions/GazeTimeoutException.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Naoray\GazeLaravel\Exceptions;
 
-class GazeTimeoutException extends GazeException
+class GazeTimeoutException extends GazeException implements TransientGazeException
 {
 }